+++ CHANGELOG.md
@@ -4,6 +4,20 @@ Dokumentasi riwayat pembaruan dan otomasi commit proyek `personal-fresh-linux-se
 
 ---
 
+### 🚀 [2026-09-04 14:12:27] - style(tui): add bigtext header, update title to Fresh Fedora KDE Utillity, and beautify banners
+
+- **Author**: Rifaul <rifaulhilal06@gmail.com>
+- **Branch**: main
+- **Files Changed**:
+  -  M fedora-setup/scripts/asus_setup.sh
+  -  M fedora-setup/scripts/flatpak_apps.sh
+  -  M fedora-setup/scripts/rpm_apps.sh
+  -  M fedora-setup/scripts/setup_terminal.sh
+  -  M fedora-setup/scripts/system_essentials.sh
+  -  M fedora-setup/setup.sh
+  - ?? bigtext
+  - ?? fedora-setup/configs/bigtext.txt
+
 ### 🚀 [2026-09-04 13:56:45] - docs: add aesthetic dev-ish README.md, changelog automation, and gsync alias
 
 - **Author**: Rifaul <rifaulhilal06@gmail.com>